Prepare for View-Model (MVVM) implementation
