+++ TVVC/Directory/Data/DirectoryItem.cs
@@ -1,0 +1,23 @@
+﻿namespace TVVC
+{
+    /// <summary>
+    /// Information about directory items, such as logical drives, files and folders
+    /// </summary>
+    public class DirectoryItem
+    {
+        /// <summary>
+        /// The type of this item
+        /// </summary>
+        public DirectoryItemType Type { get; set; }
+
+        /// <summary>
+        /// The absolute path to the item
+        /// </summary>
+        public string FullPath { get; set; }
+
+        /// <summary>
+        /// The name of the directory item
+        /// </summary>
+        public string Name { get { return this.Type == DirectoryItemType.Drive ? this.FullPath : DirectoryStructure.GetFileFolderName(this.FullPath); } }
+    }
+}
