+++ TVVC/Directory/Data/DirectoryItemType.cs
@@ -1,0 +1,12 @@
+﻿namespace TVVC
+{
+    /// <summary>
+    /// The type of a directory item
+    /// </summary>
+    public enum DirectoryItemType
+    {
+        Drive,
+        File,
+        Folder
+    }
+}
