+++ TVVC/MainWindow.xaml
@@ -4,7 +4,6 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:TVVC"
-        Loaded="Window_Loaded"
         mc:Ignorable="d"
         Title="TVVC" Height="350" Width="525">
     <Grid>
